Add new optimized function `unicode_utf8_decode`
Use it as a drop-in replacement instead of bitstring_utf8_decode.

This function is based on highly optimized UTF-8 decode found here:
http://bjoern.hoehrmann.de/utf-8/decoder/dfa/
