+++ src/libAtomVM/bitstring.h
@@ -23,6 +23,7 @@
 #define _BITSTRING_H_
 
 #include "term.h"
+#include "unicode.h"
 
 #include <stdbool.h>
 #include <stdint.h>
@@ -97,13 +98,6 @@ enum BitstringFlags
 #else
     LittleEndianIntegerMask = LittleEndianInteger
 #endif
-};
-
-enum UnicodeTransformDecodeResult
-{
-    UnicodeTransformDecodeSuccess,
-    UnicodeTransformDecodeFail,
-    UnicodeTransformDecodeIncomplete
 };
 
 union maybe_unsigned_int8
@@ -321,20 +315,6 @@ static inline bool bitstring_insert_integer(term dst_bin, size_t offset, avm_int
 bool bitstring_utf8_encode(uint32_t c, uint8_t *buf, size_t *out_size);
 
 /**
- * @brief Decode a character from UTF-8.
- *
- * @param buf the buffer from which to decode the string
- * @param len the length (in bytes) of the bytes in buf
- * @param c int value to decode to or NULL to only compute the size.
- * @param out_size the size in bytes, on output (if not NULL)
- * @return \c UnicodeTransformDecodeSuccess if decoding was successful,
- * \c UnicodeTransformDecodeFail if character starting at buf is not a valid
- * unicode character or \c UnicodeTransformDecodeIncomplete if character
- * starting at buf is a valid but incomplete transformation
- */
-enum UnicodeTransformDecodeResult bitstring_utf8_decode(const uint8_t *buf, size_t len, uint32_t *c, size_t *out_size);
-
-/**
  * @brief Encode a character to UTF-16.
  *
  * @param c character to encode
@@ -441,7 +421,7 @@ static inline bool bitstring_match_utf8(term src_bin, size_t offset, uint32_t *c
 {
     size_t byte_offset = offset >> 3; // divide by 8
     const uint8_t *src = (const uint8_t *) term_binary_data(src_bin) + byte_offset;
-    return bitstring_utf8_decode(src, term_binary_size(src_bin) - byte_offset, c, out_size) == UnicodeTransformDecodeSuccess;
+    return unicode_utf8_decode(src, term_binary_size(src_bin) - byte_offset, c, out_size) == UnicodeTransformDecodeSuccess;
 }
 
 /**
